refactor(fs): extract root-guarded relative_path helper and reuse in clean
Address PR review feedback on the clean/purge Windows path fix:

- Hoist the `current_dir()` + `dunce::canonicalize` out of `print_removing`
  into `clean_builtin`, so the cwd is canonicalized once per command instead
  of once per removed path.
- Replace the bare `pathdiff::diff_paths` call with a shared
  `pacquet_fs::relative_path`, extracted from `symlink_dir`'s
  `relative_target_inner`. It carries the Windows `same_path_root` guard,
  so a target on a different drive/UNC share renders as the absolute path
  (matching Node's `path.relative`) instead of a re-anchored garbage path.
  `symlink_dir` now delegates to the same helper rather than duplicating it.
- Move the canonicalization rationale from the `print_removing` rustdoc into
  an implementation `//` comment at the cwd computation.
