+++ pnpm/crates/cli/src/cli_args/clean.rs
@@ -4,7 +4,7 @@ use super::{
 use derive_more::{Display, Error};
 use miette::{Context, Diagnostic, IntoDiagnostic};
 use pacquet_config::Config;
-use pacquet_fs::{is_subdir, lexical_normalize};
+use pacquet_fs::{is_subdir, lexical_normalize, relative_path};
 use pacquet_workspace::read_project_manifest_only;
 use serde_json::Value;
 use std::path::{Path, PathBuf};
@@ -97,6 +97,12 @@ fn script_of(
 /// Remove `node_modules` contents and (optionally) lockfiles from the
 /// current project or every workspace project.
 fn clean_builtin(ctx: &RunCtx<'_>, config: &Config, remove_lockfile: bool) -> miette::Result<()> {
+    // The `Removing <path>` lines render each target relative to the cwd.
+    // It is canonicalized once here so it shares the symlink-resolved
+    // representation of the paths built from the canonicalized `--dir`; on
+    // Windows the raw `current_dir()` can differ (casing, 8.3 short names,
+    // junctions) and defeat the relative-path computation.
+    let cwd = std::env::current_dir().and_then(dunce::canonicalize).unwrap_or_default();
     // `pnpm clean` resolves the modules dir relative to each project
     // directory, not against a single absolute prefix, so strip the
     // config anchor back to the leaf and rejoin per project.
@@ -111,14 +117,14 @@ fn clean_builtin(ctx: &RunCtx<'_>, config: &Config, remove_lockfile: bool) -> mi
     for dir in &dirs {
         let full_modules_dir = dir.join(modules_leaf);
         if has_contents_to_remove(&full_modules_dir) {
-            print_removing(&full_modules_dir);
+            print_removing(&cwd, &full_modules_dir);
             remove_modules_dir_contents(&full_modules_dir)?;
         }
     }
     if remove_lockfile {
         let lockfile_path = root_dir.join("pnpm-lock.yaml");
         if lockfile_path.exists() {
-            print_removing(&lockfile_path);
+            print_removing(&cwd, &lockfile_path);
             std::fs::remove_file(&lockfile_path)
                 .or_else(|error| {
                     if error.kind() == std::io::ErrorKind::NotFound { Ok(()) } else { Err(error) }
@@ -140,7 +146,7 @@ fn clean_builtin(ctx: &RunCtx<'_>, config: &Config, remove_lockfile: bool) -> mi
         && is_subdir(root_dir, &resolved_virtual_store_dir)
         && resolved_virtual_store_dir.exists()
     {
-        print_removing(&resolved_virtual_store_dir);
+        print_removing(&cwd, &resolved_virtual_store_dir);
         remove_path(&resolved_virtual_store_dir)?;
     }
     Ok(())
@@ -186,18 +192,11 @@ fn is_pnpm_entry(name: &str) -> bool {
     !name.starts_with('.') || PNPM_HIDDEN_ENTRIES.contains(&name)
 }
 
-/// Print `Removing <path>` relative to the current working directory,
-/// matching pnpm's `path.relative(process.cwd(), p)` formatting. An
-/// empty relative path renders as `.`.
-///
-/// The cwd is canonicalized so it shares the symlink-resolved
-/// representation of the paths built from the canonicalized `--dir`.
-/// On Windows the raw `current_dir()` can differ from that form (casing,
-/// 8.3 short names, junctions), which would otherwise leave the two paths
-/// without a common prefix and leak the full absolute path.
-fn print_removing(path: &Path) {
-    let base = std::env::current_dir().and_then(dunce::canonicalize).unwrap_or_default();
-    let relative = pathdiff::diff_paths(path, &base).unwrap_or_else(|| path.to_path_buf());
+/// Print `Removing <path>`, with `path` rendered relative to `base` (the
+/// canonicalized cwd), matching pnpm's `path.relative(process.cwd(), p)`
+/// formatting. An empty relative path renders as `.`.
+fn print_removing(base: &Path, path: &Path) {
+    let relative = relative_path(base, path);
     let owned: PathBuf =
         if relative.as_os_str().is_empty() { PathBuf::from(".") } else { relative };
     println!("Removing {}", owned.display());
