+++ pnpm/crates/fs/src/lib.rs
@@ -1,12 +1,14 @@
 mod ensure_file;
 mod is_subdir;
 mod lexical_normalize;
+mod relative_path;
 mod symlink_dir;
 mod write_atomic;
 
 pub use ensure_file::*;
 pub use is_subdir::is_subdir;
 pub use lexical_normalize::lexical_normalize;
+pub use relative_path::relative_path;
 pub use symlink_dir::*;
 pub use write_atomic::write_atomic;
