+++ pnpm/crates/fs/src/symlink_dir.rs
@@ -71,52 +71,7 @@ fn to_native_separators(path: &Path) -> Cow<'_, Path> {
 /// two arguments.
 fn relative_target_for(original: &Path, link: &Path) -> PathBuf {
     let parent = link.parent().unwrap_or_else(|| Path::new(""));
-    relative_target_inner(original, parent)
-}
-
-#[cfg(windows)]
-fn relative_target_inner(original: &Path, parent: &Path) -> PathBuf {
-    let original = dunce::simplified(original);
-    let parent = dunce::simplified(parent);
-    if !same_path_root(original, parent) {
-        return original.to_path_buf();
-    }
-    pathdiff::diff_paths(original, parent).unwrap_or_else(|| original.to_path_buf())
-}
-
-#[cfg(not(windows))]
-fn relative_target_inner(original: &Path, parent: &Path) -> PathBuf {
-    pathdiff::diff_paths(original, parent).unwrap_or_else(|| original.to_path_buf())
-}
-
-/// Whether `a` and `b` have an identical `Component::Prefix` after
-/// `dunce::simplified`, with drive letters case-folded. UNC shares
-/// only match when their server/share are written with identical
-/// casing and variant — the check has to stay in lockstep with what
-/// `pathdiff::diff_paths` will tolerate, since a variant-tolerant or
-/// case-tolerant comparison here would let the downstream diff emit
-/// a re-anchored garbage path on a `Prefix` mismatch it cannot relate.
-#[cfg(windows)]
-fn same_path_root(a: &Path, b: &Path) -> bool {
-    fn first_prefix(path: &Path) -> Option<std::path::Prefix<'_>> {
-        match path.components().next()? {
-            std::path::Component::Prefix(p) => Some(p.kind()),
-            _ => None,
-        }
-    }
-    fn case_normalize(prefix: std::path::Prefix<'_>) -> std::path::Prefix<'_> {
-        use std::path::Prefix::{Disk, VerbatimDisk};
-        match prefix {
-            Disk(d) => Disk(d.to_ascii_uppercase()),
-            VerbatimDisk(d) => VerbatimDisk(d.to_ascii_uppercase()),
-            other => other,
-        }
-    }
-    match (first_prefix(a), first_prefix(b)) {
-        (Some(pa), Some(pb)) => case_normalize(pa) == case_normalize(pb),
-        (None, None) => true,
-        _ => false,
-    }
+    crate::relative_path(parent, original)
 }
 
 /// Remove a symlink (or junction on Windows) previously created with
